Coding_example
Hard coded Javascipt with input fields that could be filled in the Unity engine. It was used to play audio when the mouse hovered over an object that would take the player to the next stage in the the game. When clicked, the audio of cracking glass and load a new level where the snow globe the player is exploring has had it's ceiling destroyed.