Coding example (hoverscript.js)
A hard coded javascipt that would play audio on mouse hover and also on mouse click. This script was attached to particular objects scattered throughout an enormous snow globe. The script allowed for input fields that would be entered in the Unity engine so it could be reused from level to level. This script would load a new level where the roof of the snow globe had been shattered and the player could escape.